refactor(format): extract Markdown renderer to tokmd-analysis-format-md (#998)
Extract ~850 lines of Markdown rendering logic from tokmd-analysis-format
into a dedicated tokmd-analysis-format-md microcrate.

What changed:
- New crate: tokmd-analysis-format-md (pub fn render_md)
- tokmd-analysis-format now delegates to the new crate via thin wrapper
- Moved helper fns: render_file_table, render_effort_report,
  render_legacy_cocomo_report, fmt_pct, fmt_f64
- Moved corresponding unit tests to new crate
- No behavioral changes: public API (render, RenderedOutput) unchanged
- All existing tests pass; no snapshot regressions

Addresses: #998 (Phase 1 — Markdown extraction)
Next: JSON, JSON-LD, XML, SVG, Mermaid, Tree extractions (Phase 2+)
