+++ crates/tokmd-analysis-format-md/Cargo.toml
@@ -1,0 +1,22 @@
+[package]
+name = "tokmd-analysis-format-md"
+version.workspace = true
+edition.workspace = true
+license.workspace = true
+homepage.workspace = true
+repository.workspace = true
+description = "Markdown rendering for tokmd analysis receipts."
+readme = "README.md"
+keywords = ["markdown", "formatting", "code-analysis"]
+categories = ["development-tools", "encoding"]
+documentation = "https://docs.rs/tokmd-analysis-format-md"
+include = ["src/", "README.md", "Cargo.toml"]
+
+[dependencies]
+tokmd-analysis-types.workspace = true
+
+[dev-dependencies]
+insta = { workspace = true }
+tokmd-analysis.workspace = true
+tokmd-analysis-format.workspace = true
+tokmd-types.workspace = true
